add support for saved module signatures
These changes allow the user to specify input/output tensors using the
names defined in the metagraph signature for a saved model, instead of
having to know the names of the tensors in the graph itself. Tensor
names are still supported as a fallback for frozen graph (.pb) models
and existing clients. This allows us to support autograph-based models,
where tensor names can't be specified, but the names in the metagraph
signatures can.

With the additional metadata from the signatures, we also no longer
require that output tensors be specified during inference. Instead,
all output tensors in the signature are returned by default.